fixed b.controller, added test to userByID dao method

+++ src/main/java/com/budget/dao/BudgetDaoImp.java
@@ -45,7 +45,7 @@ public class BudgetDaoImp implements BudgetDao {
                 .uniqueResult();
 	}
 	
-	public User findUserById(Long id) {
+	public User findUserById(int id) {
 		return (User) sessionFactory.getCurrentSession()
                 .getNamedQuery("User.findUserById")
                 .setParameter("id", id)
